Modified CombatData with new combat properties

+++ Assets/Scripts/Entity/AttackInfo.cs
@@ -14,7 +14,7 @@
         /// <param name="damage">The amount of pain to inflict</param>
         /// <param name="damageType">The manner in which it is inflicted</param>
         /// <param name="description">A human readable description (such as "laserjet printer")</param>
-        public AttackInfo(int damage, DamageType damageType, string description) {
+        public AttackInfo(ushort damage, DamageType damageType, string description) {
             Damage = damage;
             DamageType = damageType;
             Description = description;
@@ -23,7 +23,7 @@
         /// <summary>
         ///     The damage done by this AttackInfo
         /// </summary>
-        public int Damage { get; set; }
+        public ushort Damage { get; set; }
 
         /// <summary>
         ///     The type of damage done by this AttackInfo
